Move sum-storage to super-runtime and prune api-runtime
