+++ nodes/rpc-node/Cargo.toml
@@ -20,7 +20,6 @@ parking_lot = '0.9.0'
 tokio = '0.1.22'
 trie-root = '0.15.2'
 structopt = "0.3.8"
-runtime = { package = "api-runtime", path = "../../runtimes/api-runtime" }
 jsonrpc-core = "14.0.3"
 jsonrpc-core-client = "14.0.3"
 jsonrpc-derive = "14.0.3"
@@ -46,6 +45,10 @@ sp-inherents = '2.0.0-alpha.6'
 sp-io = '2.0.0-alpha.6'
 sp-runtime = '2.0.0-alpha.6'
 sp-transaction-pool = '2.0.0-alpha.6'
+
+# This node requires its runtime to implement the sum-storage-runtime-api.
+# Only the super-runtime provides this API.
+runtime = { package = "super-runtime", path = "../../runtimes/super-runtime" }
 
 [build-dependencies]
 vergen = '3.0.4'
