--- runtimes/api-runtime/Cargo.toml
@@ -1,75 +1,0 @@
-[package]
-name = "api-runtime"
-version = "2.0.0-alpha.6"
-authors = ["Joshy Orndorff"]
-edition = "2018"
-license = "GPL-3.0-or-later"
-
-[dependencies]
-aura = { package = "pallet-aura", version = '2.0.0-alpha.6', default_features = false}
-balances = { package = "pallet-balances", version = '2.0.0-alpha.6', default_features = false}
-frame-support = { version = '2.0.0-alpha.6', default_features = false}
-grandpa = { package = "pallet-grandpa", version = '2.0.0-alpha.6', default_features = false}
-indices = { package = "pallet-indices", version = '2.0.0-alpha.6', default_features = false}
-sudo = { package = "pallet-sudo", version = '2.0.0-alpha.6', default_features = false}
-frame-system = { version = '2.0.0-alpha.6', default_features = false}
-timestamp = { package = "pallet-timestamp", version = '2.0.0-alpha.6', default_features = false}
-transaction-payment = { package = "pallet-transaction-payment", version = '2.0.0-alpha.6', default_features = false}
-randomness-collective-flip = { package = "pallet-randomness-collective-flip", version = '2.0.0-alpha.6', default_features = false}
-
-parity-scale-codec = { version = "1.3.0", default-features = false, features = ["derive"] }
-frame-executive = { version = '2.0.0-alpha.6', default_features = false}
-safe-mix = { version = "1.0.0", default-features = false }
-serde = { version = "1.0.101", optional = true, features = ["derive"] }
-sp-api = { version = '2.0.0-alpha.6', default_features = false}
-sp-block-builder = { version = '2.0.0-alpha.6', default_features = false}
-sp-consensus-aura = { version = '0.8.0-alpha.6', default_features = false}
-sp-core = { version = '2.0.0-alpha.6', default_features = false}
-sp-finality-grandpa = { version = '2.0.0-alpha.6', default_features = false}
-sp-inherents = { version = '2.0.0-alpha.6', default_features = false}
-sp-io = { version = '2.0.0-alpha.6', default_features = false}
-sp-offchain = { version = '2.0.0-alpha.6', default_features = false}
-sp-runtime = { version = '2.0.0-alpha.6', default_features = false}
-sp-session = { version = '2.0.0-alpha.6', default_features = false}
-sp-std = { version = '2.0.0-alpha.6', default_features = false}
-sp-transaction-pool = { version = '2.0.0-alpha.6', default_features = false}
-sp-version = { version = '2.0.0-alpha.6', default_features = false}
-sum-storage = { default-features = false, path = "../../pallets/sum-storage" }
-sum-storage-rpc-runtime-api = { default-features = false, path = "../../pallets/sum-storage/rpc/runtime-api" }
-
-[build-dependencies]
-wasm-builder-runner = { version = "1.0.4", package = "substrate-wasm-builder-runner" }
-
-[features]
-default = ["std"]
-std = [
-	"aura/std",
-	"balances/std",
-	"parity-scale-codec/std",
-	"frame-executive/std",
-	"frame-support/std",
-	"frame-system/std",
-	"grandpa/std",
-	"indices/std",
-	"randomness-collective-flip/std",
-	"safe-mix/std",
-	"serde",
-	"sp-api/std",
-	"sp-block-builder/std",
-	"sp-consensus-aura/std",
-	"sp-core/std",
-	"sp-finality-grandpa/std",
-	"sp-inherents/std",
-	"sp-io/std",
-	"sp-offchain/std",
-	"sp-runtime/std",
-	"sp-session/std",
-	"sp-std/std",
-	"sp-transaction-pool/std",
-	"sp-version/std",
-	"sudo/std",
-	"sum-storage/std",
-	"sum-storage-rpc-runtime-api/std",
-	"timestamp/std",
-	"transaction-payment/std",
-]
