--- runtimes/api-runtime/types.json
@@ -1,5 +1,0 @@
-{
-  "Address": "AccountId",
-  "LookupSource": "AccountId",
-  "Weight": "u32"
-}
