+++ runtimes/super-runtime/aggregate_types.js
@@ -29,6 +29,7 @@ const pallets = [
   "single-value",
   "storage-cache",
   "struct-storage",
+  "sum_storage",
   "vec-set",
 ]
